feat (sécurité): design
