+++ src/Controller/TrickController.php
@@ -26,8 +26,6 @@ final class TrickController extends AbstractController
     #[Route(name: 'list', methods: [Request::METHOD_GET])]
     public function list(): Response
     {
-        $this->addFlash('success', 'La figure a été ajoutée avec succès.');
-
         return $this->render('trick/list.html.twig');
     }
