+++ src/Twig/Component/DesignSystem/AlertComponent.php
@@ -6,7 +6,8 @@ namespace App\Twig\Component\DesignSystem;
 
 use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
 
-#[AsTwigComponent('user_nav', template: 'components/design_system/user_nav.html.twig')]
-final class UserNavComponent
+#[AsTwigComponent('alert', template: 'components/design_system/alert.html.twig')]
+final class AlertComponent
 {
+    public string $type;
 }
