+++ src/Twig/Component/Form/RegistrationComponent.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Twig\Component;
+namespace App\Twig\Component\Form;
 
 use App\Doctrine\Entity\User;
-use App\Form\ResetPasswordType;
+use App\Form\RegistrationType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Form\FormInterface;
 use Symfony\UX\LiveComponent\Attribute\AsLiveComponent;
@@ -13,17 +13,17 @@ use Symfony\UX\LiveComponent\Attribute\LiveProp;
 use Symfony\UX\LiveComponent\ComponentWithFormTrait;
 use Symfony\UX\LiveComponent\DefaultActionTrait;
 
-#[AsLiveComponent('reset_password')]
-final class ResetPasswordComponent extends AbstractController
+#[AsLiveComponent('registration', template: 'components/form/registration.html.twig')]
+final class RegistrationComponent extends AbstractController
 {
     use DefaultActionTrait;
     use ComponentWithFormTrait;
 
     #[LiveProp(fieldName: 'data')]
-    public User $user;
+    public ?User $user = null;
 
     protected function instantiateForm(): FormInterface
     {
-        return $this->createForm(ResetPasswordType::class, $this->user);
+        return $this->createForm(RegistrationType::class, $this->user);
     }
 }
