+++ src/Twig/Component/Form/ResetPasswordComponent.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Twig\Component;
+namespace App\Twig\Component\Form;
 
 use App\Doctrine\Entity\User;
-use App\Form\RegistrationType;
+use App\Form\ResetPasswordType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Form\FormInterface;
 use Symfony\UX\LiveComponent\Attribute\AsLiveComponent;
@@ -13,17 +13,17 @@ use Symfony\UX\LiveComponent\Attribute\LiveProp;
 use Symfony\UX\LiveComponent\ComponentWithFormTrait;
 use Symfony\UX\LiveComponent\DefaultActionTrait;
 
-#[AsLiveComponent('registration')]
-final class RegistrationComponent extends AbstractController
+#[AsLiveComponent('reset_password', template: 'components/form/reset_password.html.twig')]
+final class ResetPasswordComponent extends AbstractController
 {
     use DefaultActionTrait;
     use ComponentWithFormTrait;
 
     #[LiveProp(fieldName: 'data')]
-    public ?User $user = null;
+    public User $user;
 
     protected function instantiateForm(): FormInterface
     {
-        return $this->createForm(RegistrationType::class, $this->user);
+        return $this->createForm(ResetPasswordType::class, $this->user);
     }
 }
