+++ templates/components/design_system/alert.html.twig
@@ -1,0 +1,3 @@
+<div class="alert alert-{{ type }}">
+    {% block body %}{% endblock %}
+</div>
