+++ templates/components/form/reset_password.html.twig
@@ -1,0 +1,17 @@
+<div {{ attributes }}>
+    {{ form_start(form) }}
+    {{ form_row(form.plainPassword) }}
+    <div class="d-flex flex-column gap-2 align-items-center">
+        <button type="submit" class="btn btn-primary btn-lg w-100 mb-3">
+            <i class="fa-solid fa-rotate me-2"></i>
+            Réinitialiser
+        </button>
+        <p>
+            Pas encore de compte ? <a class="fw-bold" href="{{ path('security_register') }}">
+                Inscrivez-vous
+                <span class="fa-solid fa-arrow-up-right-from-square"></span>
+            </a>
+        </p>
+    </div>
+    {{ form_end(form) }}
+</div>
