--- templates/components/registration.html.twig
@@ -1,9 +1,0 @@
-<div {{ attributes }}>
-    {{ form_start(form) }}
-        {{ form_row(form.email) }}
-        {{ form_row(form.nickname) }}
-        {{ form_row(form.plainPassword) }}
-
-        <button>S'inscrire</button>
-    {{ form_end(form) }}
-</div>
