--- templates/components/request_reset_password.html.twig
@@ -1,7 +1,0 @@
-<div {{ attributes }}>
-    {{ form_start(form) }}
-        {{ form_row(form.user) }}
-
-        <button>Réinitialiser</button>
-    {{ form_end(form) }}
-</div>
