--- templates/components/reset_password.html.twig
@@ -1,7 +1,0 @@
-<div {{ attributes }}>
-    {{ form_start(form) }}
-        {{ form_row(form.plainPassword) }}
-
-        <button>Modifier</button>
-    {{ form_end(form) }}
-</div>
