--- templates/index.html.twig
@@ -1,5 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block body %}
-
-{% endblock %}
