+++ templates/security/login.html.twig
@@ -1,16 +1,57 @@
 {% extends 'base.html.twig' %}
 
 {% block body %}
-    {% if error %}
-        <div>{{ error.messageKey|trans(error.messageData, 'security') }}</div>
-    {% endif %}
-
-    <form action="{{ path('security_login') }}" method="post">
-        <input type="hidden" name="_csrf_token" value="{{ csrf_token('authenticate') }}">
-        <label for="username">Email :</label>
-        <input type="text" id="username" name="_username" value="{{ last_username }}"/>
-        <label for="password">Mot de passe :</label>
-        <input type="password" id="password" name="_password"/>
-        <button type="submit">Se connecter</button>
-    </form>
+    <div class="container flex-grow-1">
+        <div class="row justify-content-center align-items-center h-100">
+            <div class="col-12 col-md-9 col-lg-6 col-xl-5">
+                <div class="card border-0 shadow">
+                    <div class="card-body p-5">
+                        <h1 class="h3 fw-bold">Se connecter</h1>
+                        <span>Bienvenue sur Snowtricks 👋</span>
+                        <form action="{{ path('security_login') }}" method="post" class="mt-5">
+                            {% if error %}
+                                {% component alert with {type: 'danger'} %}
+                                    {% block body %}
+                                        {{ error.messageKey|trans(error.messageData, 'security') }}
+                                    {% endblock %}
+                                {% endcomponent %}
+                            {% endif %}
+                            <input type="hidden" name="_csrf_token" value="{{ csrf_token('authenticate') }}">
+                            <div class="mb-3">
+                                <label for="_username" class="form-label">Adresse email</label>
+                                <input type="email" name="_username" id="_username" class="form-control" value="{{ last_username }}">
+                            </div>
+                            <div class="mb-3">
+                                <label for="_password" class="form-label">Mot de passe</label>
+                                <input type="password" name="_password" id="_password" class="form-control">
+                            </div>
+                            <div class="mb-3 d-flex flex-column gap-2 gap-sm-0 flex-sm-row justify-content-sm-between">
+                                <div class="form-check">
+                                    <input class="form-check-input" type="checkbox" value="" id="_remember_me" checked>
+                                    <label class="form-check-label" for="_remember_me">
+                                        Se souvenir de moi
+                                    </label>
+                                </div>
+                                <a class="fw-bold" href="{{ path('security_reset_password_request') }}">
+                                    Mot de passe oublié ?
+                                </a>
+                            </div>
+                            <div class="d-flex flex-column gap-2 align-items-center">
+                                <button type="submit" class="btn btn-primary btn-lg w-100 mb-3">
+                                    <i class="fa-solid fa-right-to-bracket me-2"></i>
+                                    Se connecter
+                                </button>
+                                <p>
+                                    Pas encore de compte ? <a class="fw-bold" href="{{ path('security_register') }}">
+                                        Inscrivez-vous
+                                        <span class="fa-solid fa-arrow-up-right-from-square"></span>
+                                    </a>
+                                </p>
+                            </div>
+                        </form>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
 {% endblock %}
