+++ templates/security/register.html.twig
@@ -1,6 +1,17 @@
 {% extends 'base.html.twig' %}
 
 {% block body %}
-    {{ component('registration', {user: user.id is null ? null : user, form: form}) }}
+<div class="container flex-grow-1">
+    <div class="row justify-content-center align-items-center h-100">
+        <div class="col-12 col-md-9 col-lg-6 col-xl-5">
+            <div class="card border-0 shadow">
+                <div class="card-body p-5">
+                    <h1 class="h3 fw-bold">Réinitialisation du mot de passe</h1>
+                    {{ component('registration', {user: user.id is null ? null : user, form: form}) }}
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
 {% endblock %}
