+++ templates/security/request_reset_password.html.twig
@@ -1,9 +1,21 @@
 {% extends 'base.html.twig' %}
 
 {% block body %}
-    {{ component('request_reset_password', {
-        resetPasswordRequest: reset_password_request.id is null ? null : reset_password_request,
-        form: form
-    }) }}
+    <div class="container flex-grow-1">
+        <div class="row justify-content-center align-items-center h-100">
+            <div class="col-12 col-md-9 col-lg-6 col-xl-5">
+                <div class="card border-0 shadow">
+                    <div class="card-body p-5">
+                        <h1 class="h3 fw-bold">Mot de passe oublié ?</h1>
+                        {{ component('request_reset_password', {
+                            class: 'mt-5',
+                            resetPasswordRequest: reset_password_request.id is null ? null : reset_password_request,
+                            form: form
+                        }) }}
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
 {% endblock %}
